requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-core
langchain-embeddings
langchain-vectorstores
langchain-document-loaders
langchain-retrievers
langchain-prompts
langchain-chains
openai
flask
flask-cors
pypdf2
faiss-cpu
sentence-transformers
python-dotenv
transformers
torch